Translate each pagagraph of help text separately
With the earlier scheme, each help page was generated as one big string,
which is translated at runtime. This means that any small change in the
help text will lead to the whole page getting out of date.

Change gendocs.sh so that we generate an array of lines for each help page.
In the help display code, these lines are translated individually and then
concatenated into the page contents.

+++ docgen/gendocs.sh
@@ -29,7 +29,7 @@ echo >> ../src/help.c
 echo -n "int help_page_count = " >> ../src/help.c
 ls -l O*.doc D*.doc | wc --lines >> ../src/help.c
 echo ";" >> ../src/help.c
-echo "char *help_page_titles[] = {" >> ../src/help.c
+echo "const char *help_page_titles[] = {" >> ../src/help.c
 for A in O*.doc D*.doc
 do
 	echo -n "     $tag(\"" >> ../src/help.c
@@ -38,22 +38,35 @@ do
 done
 echo "     NULL };" >> ../src/help.c
 echo >> ../src/help.c
-echo "char *help_page_contents[] = {" >> ../src/help.c
 for A in O*.doc D*.doc
 do
+        N=`expr substr $A 1 3`
+	echo "static const char * ${N}_contents[] = {" >> ../src/help.c
 	split -l 1 $A
 	rm -f xaa
-	echo "     /* xgettext:no-c-format */ $tag(" >> ../src/help.c
 	for B in x*
 	do
-		echo -n "     \"" >> ../src/help.c
-		cat $B | sed -e "s/\\\"/\\\\\"/g" > temp.tmp
-		echo -n "`cat temp.tmp`" >> ../src/help.c
-		rm -f temp.tmp
-		echo "\\n\"" >> ../src/help.c
+		L=`cat $B`
+		if test "x$L" = "x"
+		then
+			echo    "     \"\\n\"," >> ../src/help.c
+		else
+	        	echo -n "     /* xgettext:no-c-format */ $tag(\"" >> ../src/help.c
+			cat $B | sed -e "s/\\\"/\\\\\"/g" > temp.tmp
+			echo -n "`cat temp.tmp`" >> ../src/help.c
+			rm -f temp.tmp
+			echo "\\n\")," >> ../src/help.c
+		fi
 	done
 	rm -f x*
-	echo "     )," >> ../src/help.c
+	echo "     NULL };" >> ../src/help.c
+        echo >> ../src/help.c
+done
+echo "const char ** help_page_contents[] = {" >> ../src/help.c
+for A in O*.doc D*.doc
+do
+        N=`expr substr $A 1 3`
+	echo "     ${N}_contents ," >> ../src/help.c
 done
 echo "     NULL };" >> ../src/help.c
